Fixed formatting and removed MaterialType() from myr api
