+++ src/index.js
@@ -1,6 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
+import "./utils/AFrameMaterials";
 import "./utils/AFramePhysics";
 import "./utils/WASDPlusControls";
 import "./utils/AframeRegIndicator";
